Change kAshmemMemoryLimit to int64
R=reveman
BUG=462348

[Android] Fix BufferUnderflowException in webrtc buffer

The image buffer may not contain all bytes in the last row, causing an
exception. Fix by only reading up to what's remaining() in the buffer.

BUG=458701

Review URL: https://codereview.chromium.org/975913003

Cr-Commit-Position: refs/heads/master@{#319401}
(cherry picked from commit e64698bbd6c20bfdad48db99d65a04c25fb02834)

Review URL: https://codereview.chromium.org/1012793004

Cr-Commit-Position: refs/branch-heads/2272_90@{#5}
Cr-Branched-From: a4600628d9d127b3e8aed98d51c25158ddd2267f-refs/branch-heads/2272@{#433}
Cr-Branched-From: 827a380cfdb31aa54c8d56e63ce2c3fd8c3ba4d4-refs/heads/master@{#310958}

+++ base/memory/discardable_memory_android.cc
@@ -23,7 +23,7 @@ const char kAshmemAllocatorName[] = "DiscardableMemoryAshmemAllocator";
 
 // For Ashmem, have the DiscardableMemoryManager trigger userspace eviction
 // when address space usage gets too high (e.g. 512 MBytes).
-const size_t kAshmemMemoryLimit = 512 * 1024 * 1024;
+const int64 kAshmemMemoryLimit = 512 * 1024 * 1024;
 
 size_t GetAshmemMemoryLimit() {
   // Allow 25% of physical memory to be used for discardable memory.
